chore: add @tailor-ui/utils
