+++ doczrc.js
@@ -15,6 +15,7 @@ module.exports = {
         alias: {
           'tailor-ui': './packages/tailor-ui/src',
           '@tailor-ui/lab': './packages/tailor-ui-lab/src',
+          '@tailor-ui/utils': './packages/tailor-ui-utils/src',
         },
       },
     ]);
