+++ packages/tailor-ui-utils/src/index.ts
@@ -1,0 +1,1 @@
+export { default as mergeEventProps } from './mergeEventProps';
