+++ packages/tailor-ui-utils/tsconfig.json
@@ -2,7 +2,6 @@
   "include": ["src/**/*"],
   "exclude": ["**/__tests__"],
   "extends": "../../tsconfig",
-  "references": [{ "path": "../tailor-ui" }, { "path": "../tailor-ui-utils" }],
   "compilerOptions": {
     "outDir": "./lib" /* Redirect output structure to the directory. */,
     "rootDir": "./src" /* Specify the root dir */
